Fix order of charge versions in view licence page (#1162)
https://eaflood.atlassian.net/browse/WATER-4556

> Part of the work to replace the legacy view licence page

Spotted in testing is that if a licence has 2 charge versions with the same start date (one will be 'approved' and the other will display as 'replaced'), it is possible that they will be ordered incorrectly. In this scenario, the 'approved' should be first in the table (the one without an end date) and then the 'replaced'.

Because we only sort by start date, when we encounter 2 charge versions with the same start date, we are not in control of the order they are displayed in.

This change fixes the issue by simply adding sort by end date `DESC` to the query.