Fix unit tests after enablingSystemLicenceView tag
https://eaflood.atlassian.net/browse/WATER-4555

As the new licence view page for `water-abstraction-system` is now being enabled, this has caused some of our unit tests to fail.
After looking at some of the failing tests, this seems to be because of a new URL endpoint being used.

This PR will fix the failing unit tests within `water-abstraction-system`.