Make Bill run value blank when in Review Status (#1164)
https://eaflood.atlassian.net/browse/WATER-4549

During the testing for the two-part tariff review pages, it was noted by the billing and data team that when a 2pt bill run is displayed with a status of review then the value column is always £0.00. The business requested that this column be empty rather than showing a figure of £0.00. This PR makes that change.