Fix order of charge vers. in view licence - again! (#1168)
https://eaflood.atlassian.net/browse/WATER-4556

> Part of the work to replace the legacy view licence page

We thought we'd nailed the ordering of the charge versions in the new licence setup tab of our new view licence screen in [Fix order of charge versions in view licence page](https://github.com/DEFRA/water-abstraction-system/pull/1162).

However, it's still not quite right. The previous fix assumed a 'superseded' charge version would always have an end date. However, it turns out this is not always the case. If followed by a 'current' charge version that also has an end date (because another has been added after it), sorting by end date means the 'current' comes after the 'superseded'.

Having double-checked the legacy code, they are sorting by start date, then version number. As it is the legacy view's order we're trying to match, we're just going to crack on and make the change to do the same thing.