Spike migrating import licence to system
https://eaflood.atlassian.net/jira/software/c/projects/WATER/boards/96?quickFilter=347&selectedIssue=WATER-4536

We need to explore and establish how we could migrate the import licence process from the import service.

Questions have been raised around packages used by the import service and wether or not we will keep them or rewrite that functionality. For example, we will replace moment.js with node date time.

A mural board has been made to capture the process, dependencies and order in which we would need to do this work.