Update presenter tests
We do a little housekeeping to bring the presenter's tests a little closer to our conventions and deal with some linting issues. We also make the starting data more realistic by aligning the dates and not populating te end dates.

Beyond that we change the assertions in the tests to match how we expect the data to be displayed in the page.