Fix dates in view licence set up tab (#1175)
https://eaflood.atlassian.net/browse/WATER-4556

> Part of the work to replace the legacy view licence page

In testing of the new view licence page, specifically the licence set up tab, it was spotted that when a charge version has a status of `REVIEW` that the date the record was created rather than the start date is displayed in the charge information table.

This is because the record is not a charge version; it is a workflow record, which, when approved, will become a charge version record. The current presenter logic is not distinguishing between the different types of workflow records, hence the wrong date being displayed.

But hold on. A quick review of the legacy version of the tab shows that when a workflow record does not have a status of `REVIEW` we don't show any dates at all!

Then we spotted that there is inconsistency about whether we use a dash in the end date column. The charge information table does, but the return requirements and agreements tables don't.

So, this change is a general fix-up of dates in the tab.

- for workflow records in `REVIEW` show the start date from the stored charge version details
- for all tables where a date is not set, stop displaying `-`
- do not display any dates for workflow records that have a status other than `REVIEW`