Fix workflow start date in view licence set up tab
https://eaflood.atlassian.net/browse/WATER-4556

In testing of the new view licence page, specifically the licence set up tab, it was spotted that when a charge version has a status of `REVIEW` that the date the record was created rather than the start date is displayed in the charge information table.

We fixed that in [Fix dates in view licence set up tab](https://github.com/DEFRA/water-abstraction-system/pull/1175). But hot on its heels someone has pointed out the same applies for a workflow record with a status of `changes_requested`.

So, this change fixes it so we display the charge version start date for a workflow record with a status of `changes_requested` in the charge information table of the Licence setup tab in our new view licence page.