Fix workflow start date in view licence set up tab (#1176)
https://eaflood.atlassian.net/browse/WATER-4556

In testing of the new view licence page, specifically the licence set up tab, it was spotted that when a charge version has a status of `REVIEW` that the date the record was created rather than the start date is displayed in the charge information table.

We fixed that in [Fix dates in view licence set up tab](https://github.com/DEFRA/water-abstraction-system/pull/1175). But hot on its heels, someone has pointed out that the same applies for a workflow record with a status of `changes_requested`.

So, this change fixes it, so we display the charge version start date for a workflow record with a status of `changes_requested` in the charge information table of the Licence setup tab in our new view licence page.

---

Because this meant looking at another workflow status, we needed to check that everything was linked up for `changes_requested`. This led us to take another look at how the status for each record type in the tab was being determined. We decided to split the types to give us more flexibility and use the same convention we've used in other tags, i.e. pass the record status through and let the macro handle the display.

So, a change to our tag macros was also made in this change.