Update SubmitPurposeService
This is the core of 2 changes. Needing to combine the aliases and the selected purposes from the request. But also to start persisting the description to save us having to fetch it again in the check page.

We also need to handle the edge case of a user entering a description that is too long. Before the only validation issue was that they hadn't selected anything so there was nothing to play back.

Now with the alias they may have selected some purposes but also caused a validation error. We need to ensure what they submitted is played back to them.