Fix duplicate purposes in return reqs. set up
https://eaflood.atlassian.net/browse/WATER-4592

> Part of the return requirements set up work

During the return requirements set up journey, users need to select at least one purpose for each return requirement they are setting up. The purposes they can choose from are only those linked to the current version of the licence.

The issue found is that if a licence is linked to a purpose and then to a point, then the same purpose again but to a different point (perhaps because it has condition or different abstraction properties), it causes our logic to return the same purpose twice. More if things are really complicated!

This change fixes it so that the user only sees a distinct list of applicable purposes on the 'Select the purpose for the requirements for returns' page.