Update display of points in return requirements (#1171)
* Update display of points in return requirements

https://eaflood.atlassian.net/browse/WATER-4567

Currently, when we display the points for a return requirement in both the /check page and view page we simply list them. However, we know there are a
number of licences which can be linked to more than 5 points. The most extreme is 36! It is a common pattern in the service that when we need to display
more than 5 items, we’ll use a [GOV.UK details component](https://design-system.service.gov.uk/components/details/).

This PR is focused on amending both return requirement views to use a details component if a return requirement is linked to more than 5 points. If the requirement
is linked to 5 or less, the points should be listed out (no change).