Display the purpose desc (alias) in rtn req view
https://eaflood.atlassian.net/browse/WATER-4582

> Part of the return requirements set up work

When setting up new return requirements, you are required to select one or more purposes. By default, we use the purpose description. But because the returns are intended to reflect what is on the licence document, sometimes there is a discrepancy between how the system describes the purpose and how it was recorded in the document.

To deal with this, NALD has the concept of a 'purpose alias'. This is an optional alternate description that should be used when referring to the purpose on a return requirement.

This change updates the return requirements page to display the purpose description (alias).