Name fetch properties after entities
The properties were renamed to make them less verbose. However, in all other fetch services based on models and their relationships we have stuck with the relationship names. Working on the principle of least surprise we're taking the opportunity to refactor the names to follow the convention of other services.