Only use 1st site description in return reqs set up (#1184)
https://eaflood.atlassian.net/browse/WATER-4442

> Part of the work to replace NALD for handling return requirements

In [Use abstraction data to create return requirements](https://github.com/DEFRA/water-abstraction-system/pull/1107) we added the logic to the return requirements set up journey that allows a user to generate return requirements from a licence's existing abstraction data.

One of the scenarios we had to tackle was how to populate the site description for a return requirement that has multiple points. At the time, we went with concatenating them all using ' - '.

After some additional testing, it has been decided that it is clearer if we just use the first available site description. Whatever we opt for we can see users will change it anyway. Just picking one at least improves how the return requirement is displayed on the page.