Update ret. req. setup use abstraction & copy
https://eaflood.atlassian.net/browse/WATER-4573

> Part of the return requirements set up work

In [Add optional alias to return requirement purpose](https://github.com/DEFRA/water-abstraction-system/pull/1177) we added the ability to assign a purpose description (alias) to the purposes selected when setting up a new return requirement.

As part of that we made a change to how we store the purposes in the session. No longer do we just store the ID, we now store an object.

```javascript
{
  purposes: [{
    alias: 'spray all over the place',
    description: 'Spray Irrigation - Direct',
    id: '42d5f68d-0b7a-4d29-b04e-943362746f59'
  }]
}
```

The 'Select a purpose' now expects the purposes inside the session to be in this format, and will persist them in this manner when you submit your selection. But the 'Start with abstraction data' and 'Copy existing requirement' journeys are still persisting the purposes in the session using the old `{ purposes: ['42d5f68d-0b7a-4d29-b04e-943362746f59'] }` format.

This change updates them and gets all 3 journeys working again.