Add bookmarks to return reqs check page
https://eaflood.atlassian.net/browse/WATER-4593

> Part of the return requirements set up work

When setting up return requirements for a licence it is possible that you need quite a few! Some licences have more than 10. Because of the amount of information needed in each, it makes the `/check` page hard to navigate.

To help with this we're adding bookmarks to the page that will allow users to quickly navigate to each one. The bookmarks will only display when there is more than one.

That's it!