Fix error in generate 2PT bill run transactions
https://eaflood.atlassian.net/browse/WATER-4196

> Part of the two-part tariff annual billing work

We recently completed [implementation to the annual two-part tariff billing engine](https://github.com/DEFRA/water-abstraction-system/pull/1172). But we've immediately spotted an error with the results.

We are not populating the correct field in order to see the amended billing volume being sent to the Charging Module and persisted against the transaction.

This change corrects the issue.