Fix error in generate 2PT bill run transactions (#1188)
https://eaflood.atlassian.net/browse/WATER-4196

> Part of the two-part tariff annual billing work

We recently completed [implementation of the annual two-part tariff billing engine](https://github.com/DEFRA/water-abstraction-system/pull/1172). But we've immediately spotted an error with the results.

We are not populating the correct field in order to see the amended billing volume being sent to the Charging Module and persisting against the transaction.

However, that uncovered something else we'd overlooked. Now that we were using the review allocated amount, some transactions were being returned from `GenerateTransactionService` with a 0 volume, which caused the Charging Module to error.

Of course! If a return was submitted as a 'Nil return' then we should not be trying to generate a charge for those charge elements linked to it. 🤦 

So, this change corrects which value we are assigning to the generated transaction and therefore passing to the charging module. But it also updates the engine to handle ignoring transactions with no volume rather than attempting to send them to the Charging Module API.