Alter two-part tariff returns allocation by date
https://eaflood.atlassian.net/browse/WATER-4586

During testing for the two-part tariff engine it was raised that some returns when split over two charge versions where allocating incorrectly to the first charge version rather than the second one. The example given is a licence with two charge versions. the first charge version ends on the 16th of May and the second one begins on the 17th. The returns are done on a monthly frequency, so the first return which ended in April correctly allocated to the first charge element. However the second return which starts with the month of May, allocated the first line (01/05/22 - 31/05/22), to the first charge element again, rather than the second one. The business would like it that when a return has a frequency of monthly or weekly and it is split over a charge versio, that the return volume for the week/month is allocated to the charge version that is live at the month end or week end. This is PR is changing the way we deal with allocating the lines. We have added in an additional date check where a lines end date must be within the end date of the abstraction period its trying to allocate to.