create wrapper for licence-version-purpose-condition table (#1181)
As part of the work to replace the import service we need to be sure that the data that we are inserting into the database is valid.

This PR creates a wrapper for the licence-version-purpose-condition table. It will validate the data being inserted or updated.

This table also uses some reference data from the licence-version-purpose-condition-type table. This PR adds a seeder to push that data into the unit test database and updates the tests that used the previous helper to use the new database wrapper.