Add region test seeder
https://eaflood.atlassian.net/browse/WATER-4575

The public.regions table is what we call reference data. We assume this table will not change frequently, if ever.

As part of the work to replace the legacy import service we have decided to seed any data we have deemed reference data.

We will use these seeders in testing. We need to add this now as the nald region id when randomly generated is between 1-8. This was causing random test failures when using the normal RegionHelper.add().

This was found because we have moved away from clearing the database before each test and having a pre test clean the database instead. (This is previous work to allow us to potentially speed up the tests and change our test framework)