Import Licence  versions
We need to replace the import service logic to import a licence from NALD.

We will leave the import service running to add the NALD data in the import tables, but we will slowly replace the import-licence functionality and move this into the system repo.

https://eaflood.atlassian.net/browse/WATER-4575

This change will add the licence versions to the import logic.

The licence versions rely on the created / existing licence id.

This work needs to be done first to allow us to remove the licence and licence versions from the old import logic.