Update /workflow page to remove 'charge' elements
https://eaflood.atlassian.net/browse/WATER-4570

With WRLS taking over return information when a licence is added or changed and a record is added to workflow,
the workflow record needs to direct users to check both return and charge information, not just charge information.
It also needs to stop referring to ‘charge information’ only, for example, in the link on the manage page and on the page itself.

This PR is focused on amending the /manage page and it's subsections for the new licence set up view.