Two-part tariff due returns to show as blank
https://eaflood.atlassian.net/browse/WATER-4606

During testing for the two-part tariff review pages, it was raised that the billing and data team wanted to amend the returns displaying under a charge element. Currently when a return matches to a charge element, we show these under the charge element with there allocated quantity. This means for a return that is due the quantity displayed is always 0. The billing and data team would like to remove this and show a due return as being blank instead of 0. They say that showing 0 gives the false impression that the return is a nil return, rather then a due return. This PR is for this change.