Amend licence holder display two-part tariff review pages
https://eaflood.atlassian.net/browse/WATER-4537

During testing for the two-part tariff review pages, it was decided that we would amend the licence holder display. To make it clearer the licence holder name has been moved to the top of the licence display and the billing account link would be updated to include the account name. This PR is for those changes.