Amend licence holder display two-part tariff review pages (#1199)
https://eaflood.atlassian.net/browse/WATER-4537

During testing for the two-part tariff review pages, we decided to amend the licence holder display. The licence holder's name has been moved to the top of the licence display and the billing account link will be updated to include the account name. This makes the page cleaner and avoids duplication with the licence holder's name being displayed. This PR is for those changes.