Two-part tariff void returns to show as void on review screens
https://eaflood.atlassian.net/browse/WATER-4606

During testing for the two-part tariff review pages, it was raised that the billing and data team wanted to amend the returns displaying under a charge element. When a return mathes a charge element, we show these under the review charge element section with their allocated quantity. A void return currently displays 0 ML. The billing and data team wants to remove this and show a void return as "void" instead of 0 ML. They say that showing 0 ML gives the false impression that the return is a nil return, rather than a void return. This PR is for this change.