https://eaflood.atlassian.net/browse/WATER-4579
The public.purposes table is what we call reference data. We assume this table will not change frequently, if ever.

As part of the work to replace the legacy import service we have decided to seed any data we have deemed reference data.

We will use these seeders in testing.

This change will add a seeder for the reference data for purposes and export its data to be used in tests. The id's will be fixed.