Some returns not received error incorrectly applies 2pt (#1200)
https://eaflood.atlassian.net/browse/WATER-4605

During testing for the two-part tariff review pages, the issue of Some returns not received has been picked up as being incorrectly used. When the status of a return is due, an issue flag gets marked against that return saying 'No returns received.' If this return matches a charge element an additional flag gets marked on the charge element saying 'Some returns not received'. Whilst technically this is correct based on the initial requirements, the business has since asked for this to be changed.

The changes are as follows:

For a return with a status of due, that matches a charge element and the charge element only matches the one return, then the issue 'No returns received' will be persisted against the return AND the charge element. The charge element will not have the issue 'Some returns not received' (because it only matched one return that wasn't received, the text 'some' here does not make sense.)The implications of this mean that when all the issues on the returns and the charge elements are added together to make a list for the licence because it is the same issue on the return and charge element it will be deduplicated and therefore the licence won't get additionally marked as having 'Multiple issues'. Another implication is when a user filters the licence by 'No returns received' they will see a list of that issue from the return and charge elements. However the business is ok with this, and it would likely not cause any issues.

For a return with a status of due, that matched a charge element and the charge element has also matched other returns, then the issue 'No returns received' will be flagged against the return and the current issue of 'Some returns not received' will continue to flag on the charge element. This is no different to the current behaviour.

Lastly, for a charge element that has matched multiple returns and all the returns have a status of due, this will be like scenario one again where the return and the charge element get flagged as 'No returns received'.