Update following rebase
We've rebased against `main`` after a long period of time and there have been many updates since we started this change.

This puts back in some changes we following the merge, and adds the crumb input to the additional pages in the journey that didn't exist when we were trialling this.