Solve problem of how to test crumb enabled endpoints
After much debugging we were able to see that crumb generates a token it then stores in a cookie during the `onPreResponse` hook.

Assuming we have added `<input type="hidden" name="wrlsCrumb" value="{{wrlsCrumb}}"/>` to the form and not disabled crumb in the route options for the endpoint, the cookie is included in the headers and the `wrlsCrumb` in the payload.

Hapi extracts the cookie and makes it available on the request object, from which crumb the plugin can retrieve it. It compares this value to `wrlsCrumb` in the payload.

If they match it deletes `wrlsCrumb` from the payload (hence you don't see it in the controller's version of the payload), else it causes a not authorised (and in our setup this means a 404).

So, as mentioned we saw that a number of the Defra `ffc-*` projects that use crumb share the same util function that fires a GET request, retrieves the crumb token and returns it.

The tests then include this value in the request headers as a cookie and in the payload. Our testing confirmed you got the same result (a validated crumb token) by just doing the latter! There is no need to generate a token using a GET request first.

As we'll need to do this for a number of tests, and knowing we already apply the same auth options across the majority of POST requests, we came up with this helper that we intend to switch the tests to as we update them to work with crumb.