Persist a new return version for 'no returns required' https://eaflood.atlassian.net/browse/WATER-4619
Having completed the return requirements setup journey (no-returns-required), we need to persist a return version (start date, reason and notes) to the database. There will be no return requirements in this scenario.

When "Approve" is clicked in the `/check` page the data in the session needs to be persisted to the `water.return_versions` table.