Amend reasons for 'No returns needed' journey
https://eaflood.atlassian.net/browse/WATER-4612

A decision has been made to change the options on the 'no returns needed' page during
the no returns needed journey. This also includes changing any licences with a reason
of 'Transfer licence' must be changed to the new option of 'Licence conditions do not
require returns'.

This PR is focused on amending the option from 'Transfer licence' to 'Licence conditions do not
require returns' and updating the database to match these changes.