Update existing return versions when a new one is added https://eaflood.atlassian.net/browse/WATER-4568
Having worked out how to convert the information in the return requirements setup session into the various records in the DB (see https://github.com/DEFRA/water-abstraction-system/pull/1137) we now need to deal with how the new record impacts the existing ones.

This PR will extend the persisting logic to deal with changes to the existing return versions.