Refactor the existing purpose condition types to be consistent (#1218)
* Refactor the existing purpose condition types to be consistent

As part of migrating the import service into the wabs system repo we are adding seeders to the tests which will add what we call reference data the test database.

This is currently running and adding other seeders in the test database.

This change aligns the licence version purpose condition type seeder. Mainly using a data file for us to access this static data in the tests and using the build value string function to build the values to insert into the database.

https://eaflood.atlassian.net/browse/WATER-4535

* chore: change log to error