Deleted agreements shouldn't show in the setup tab
https://eaflood.atlassian.net/browse/WATER-4604

Currently, when an agreement is deleted within the licence set up tab, the agreement still remains.
This is because the agreement is being 'soft deleted' rather than being removed from the database
completely by adding the current time and date to the 'deleted_at' column in the licence agreements table.
The logic for pulling licence agreements is missing instructions to only show licence agreements where the 'deleted_at' is null.

This PR is focused on not showing licence agreements in the licence set up tab where the licence has the 'deleted_at' column populated.