Change 'Transfer licence' reason for 'No returns needed' journey (#1217)
* Change 'Transfer licence' reason for 'No returns needed' journey

https://eaflood.atlassian.net/browse/WATER-4612

A decision has been made to change the 'Transfer licence' option for the 'no returns needed' page to 'Licence conditions do not require returns'. This also includes changing the return version reason being saved in the database as 'transfer-licence' to 'licence-conditions-do-not-require-returns'.

This PR is focused on amending the option from 'Transfer licence' to 'Licence conditions do not require returns'.