Adding CSRF form protection via Hapi Crumb plugin (#911)
https://github.com/DEFRA/water-abstraction-team/issues/114

[Crumb](https://hapi.dev/module/crumb/) is used to diminish CSRF attacks using a random unique token that is validated on the server side. **Crumb** may be used whenever you want to prevent malicious code from executing system commands that are performed by HTTP requests.

After this change, any view that has a form that uses `method="POST"`, will need to have a hidden input field.

```html
  <input type="hidden" name="wrlsCrumb" value="{{wrlsCrumb}}"/>
```

When the page is requested, the **Crumb** plugin will generate a token in the `onPreResponse` event that it will save to a cookie called 'wrlsCrumb'. It also makes it available in the view context, hence our views can reference `{{wrlsCrumb}}`.

When the user submits the form, the **Crumb** plugin jumps in again, this time on the `onPostAuth` event. It compares both values, and if they match, it 'authorises' the request.

If the payload is missing `{{wrlsCrumb}}`, or the value doesn't match that saved in the cookie (which is secure and unreadable to the client) then it rejects the request. In our service, the user will see a 404, as that is the default for an 'unauthorised' request.

**Crumb** will be enabled by default for all endpoints to avoid us forgetting to protect an HTML form. However, this means it performs the check for _all_ POST requests. Our service exposes API-only POST endpoints, which do not expect a payload. For this, the route config must include the following to disable **Crumb**.

```javascript
 options: {
   plugins: {
     crumb: false
   }
 }
```