Return version can be on/after licence start date
https://eaflood.atlassian.net/browse/WATER-4599

Currently, when a user reaches the start date page for either the 'no returns required' or
'set up new requirements' journey, and enters a start date before the return version start date, the system errors.
A fix needs to be put in place to allow users to enter a start date that is on or after the licence start date.

This PR will change the validation to allow users to enter a start that can proceed the return version start date
but is on or after the licence start date.