Returns setup options work for 'dead' licences
https://eaflood.atlassian.net/browse/WATER-4600

Currently during the returns set up journey when a licence is lapsed, revoked, or expired, selecting the options
of ‘Start by using abstraction data’, ‘Copy from existing requirements’ and ‘Set up manually’ does not work. This needs to work
as there are instances were historic corrections may need to be applied to even a licence which is 'dead'.

This PR is focused on allowing licences that are 'dead' to continue after selecting any of the options on the set up page and complete the returns journey.