Add insert do update on conflict to region seeder
https://eaflood.atlassian.net/browse/WATER-4535

As we move to using seeders for tests we can look to add functionality to allow us to seed the actually database.

This change adds ON CONLICT resoultion based on unique constraints that are not the primary id's. This will allow the seeders to seed data and not affect existing data.

This change is part of ongoing work to with the import service and test improvements to allow us to move to parallel test runs.

Regions does not have unique constraint on the  charge_region_id or the nald_region_id so we will either update where they match or insert where the pair does not exist.